fix behavior on windows
This is (or at least, feels like) a combination of two hacks.
Notably, these hacks seem to apply only when using `pyproject.toml` files.
The multiple inputs test that uses only `requirements.in` files passes on
Windows even without this change.

First, `pip-compile` fails when using the the default (`setuptools`).
The error is:
```
Backend subprocess exited when trying to invoke get_requires_for_build_wheel
Failed to parse <PATH_TO_RUNFILES_DIR>\runfiles\rules_python\tests\multiple_inputs\a\pyproject.toml
pip-compile exited with code 2. This means that pip-compile found incompatible requirements or could not find a version that matches the install requirement in one of ['tests/multiple_inputs/a/pyproject.toml', 'tests/multiple_inputs/b/pyproject.toml'].
```
To work around this, I have changed the `build-system` used in the `pyproject.toml` files to `hatchling`.

Second, it seems that package names from pyproject.toml files are not treated
differently on different platforms. To work around this, we set `requirements_windows` to
have separate lockfiles for different platforms.
Perhaps package names in the output should be normalized (according to https://packaging.python.org/en/latest/specifications/name-normalization/).
I'm unsure if this change would need to be made upstream (in `pip-tools`) or if it's just a `rules_python` thing.

It's unclear to me what (if any) of this is known/intentional, or if I've stumbled into some
odd behavior that ought to be fixed.